MDEV-18851: factor large page initialization into my_init_large_pages
